Move db functions used to write to EventStore
